backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
Pillow==10.4.0
google-genai==1.5.0
httpx==0.27.0
python-dotenv==1.0.1
reportlab==4.4.10
